Move task crud operations to its own cog
